Decoupling after refactoring
