+++ src/main/java/com/mycompany/knockknock/KnockKnock.java
@@ -21,52 +21,13 @@ public class KnockKnock {
      */
     public static void main( String[] args ) {
 
-        final Exchanger<String> exchanger = new Exchanger<String>();
+        final Exchanger<String> exchanger = new Exchanger<>();
 
-        Thread person1 = new Thread(new Runnable() {
+        Joe joe = new Joe("Joe", exchanger);
+        Alice alice = new Alice("Alice", exchanger);
 
-            @Override
-            public void run() {
-                try {
-                    String reply = exchanger.exchange("Knock Knock");
-                    System.out.println(reply);
-
-                    reply = exchanger.exchange("Dozen");
-                    System.out.println(reply);
-
-                    reply = exchanger.exchange("Doesn't anybody want to let me in!");
-
-
-                } catch (InterruptedException e) {
-                    System.err.println(e);
-                }
-            }
-        });
-
-
-        Thread person2 = new Thread(new Runnable() {
-            @Override
-            public void run() {
-                try {
-                    String reply = exchanger.exchange("Who's there?");
-                    System.out.println(reply);
-
-                    reply = exchanger.exchange("Dozen" + " who?");
-                    System.out.println(reply);
-
-                    reply = exchanger.exchange("");
-                    System.out.println(reply);
-
-
-                } catch (InterruptedException e) {
-                    System.err.println(e);
-                }
-            }
-        });
-
-        person2.start();
-        person1.start();
-
+        new Thread(joe).start();
+        new Thread(alice).start();
 
     }
 }
