--- src/main/java/com/mycompany/knockknock/Person.java
@@ -1,10 +1,0 @@
-/*
- * |-------------------------------------------------
- * | Copyright © 2016 Colin But. All rights reserved. 
- * |-------------------------------------------------
- */
-package com.mycompany.knockknock;
-
-public class Person {
-
-}
